+++ mobileapp/adapters/lib/controllers.dart
@@ -35,7 +35,11 @@ class ApplicationWideController {
 
   final KnowledgebaseUseCases _knowledgebaseUsecases = KnowledgebaseUseCases(DependencyProvider());
 
-  final RouteDbUseCases _routeDbUseCases = RouteDbUseCases(DependencyProvider());
+  /// The use case object from the `core` ring.
+  final RouteDbUseCases _routeDbUseCases;
+
+  /// Constructor for creating a new instance.
+  ApplicationWideController() : _routeDbUseCases = DependencyProvider().provide<RouteDbUseCases>();
 
   /// The user requested a switch to the Journal domain.
   void requestSwitchToJournal() {
@@ -89,7 +93,10 @@ class KnowledgebaseController {
 /// Controller for transmitting route DB UI messages to the core.
 class RouteDbController {
   /// The use case object from the `core` ring.
-  final RouteDbUseCases _routeDbUseCases = RouteDbUseCases(DependencyProvider());
+  final RouteDbUseCases _routeDbUseCases;
+
+  /// Constructor for creating a new instance.
+  RouteDbController() : _routeDbUseCases = DependencyProvider().provide<RouteDbUseCases>();
 
   /// The user requested updating the route database.
   void requestRouteDbUpdate() {
@@ -104,9 +111,9 @@ class RouteDbController {
   }
 
   /// The user requested to filter the summit list by [filterText].
-  void requestFilterSummitList(String filterText) {
-    _logger.debug('UI request: Filter summit list for "$filterText"');
-    unawaited(_routeDbUseCases.filterSummitList(filterText));
+  void requestFilterSummitList(String filterText, ItemDataId? filterArea) {
+    _logger.debug('UI request: Filter summit list for "$filterText" in area: $filterArea');
+    unawaited(_routeDbUseCases.filterSummitList(filterText, filterArea));
   }
 
   /// The user requested to see all details of the single summit identified by [summitDataId].