+++ mobileapp/adapters/lib/storage/routedb.dart
@@ -10,6 +10,7 @@ import 'package:core/entities/data_source.dart';
 import 'package:core/entities/geoposition.dart';
 import 'package:core/entities/post.dart';
 import 'package:core/entities/route.dart';
+import 'package:core/entities/sector.dart';
 import 'package:core/entities/sorting/posts_filter_mode.dart';
 import 'package:core/entities/sorting/routes_filter_mode.dart';
 import 'package:core/entities/summit.dart';
@@ -228,6 +229,29 @@ class RouteDbStorage implements RouteDbStorageBoundary {
   }
 
   @override
+  Future<List<Sector>> retrieveAllSectors() async {
+    // Configure the query
+    Query query = Query.table(
+      AreasTable.tableName,
+      <String>[AreasTable.columnId, AreasTable.columnName],
+    );
+
+    query.orderByColumns = <String>[AreasTable.columnName];
+
+    // Run the database query
+    List<ResultRow> resultSet = await _repository.executeQuery(query);
+
+    // Convert and return the result data
+    List<Sector> sectors = <Sector>[];
+    for (final ResultRow dataRow in resultSet) {
+      int id = dataRow.getIntValue(AreasTable.columnId);
+      String name = dataRow.getStringValue(AreasTable.columnName);
+      sectors.add(Sector(id, name));
+    }
+    return sectors;
+  }
+
+  @override
   Future<Summit> retrieveSummit(int summitDataId) async {
     Query query = Query.join(
       <String>[SummitsTable.tableName, SummitNamesTable.tableName, AreasTable.tableName],
@@ -256,7 +280,7 @@ class RouteDbStorage implements RouteDbStorageBoundary {
   }
 
   @override
-  Future<List<Summit>> retrieveSummits([String? nameFilter]) async {
+  Future<List<Summit>> retrieveSummits([String? nameFilter, int? sectorIdFilter]) async {
     // Configure the query
     Query query = Query.join(
       <String>[SummitsTable.tableName, SummitNamesTable.tableName, AreasTable.tableName],
@@ -273,12 +297,25 @@ class RouteDbStorage implements RouteDbStorageBoundary {
       ],
     );
 
-    if (nameFilter != null) {
-      _logger.debug('Retrieving filtered summit list for "$nameFilter"');
-      query.setWhereCondition('${SummitNamesTable.columnName} LIKE ?', <String>['%$nameFilter%']);
+    List<String> whereClauses = <String>[];
+    List<Object> whereArgs = <Object>[];
+
+    if (nameFilter != null || sectorIdFilter != null) {
+      if (nameFilter != null) {
+        _logger.debug('Filtering summit list summit list for "$nameFilter"');
+        whereClauses.add('${SummitNamesTable.columnName} LIKE ?');
+        whereArgs.add('%$nameFilter%');
+      }
+      if (sectorIdFilter != null) {
+        _logger.debug('Filtering summit list for sector $sectorIdFilter');
+        whereClauses.add('${AreasTable.columnId} = ?');
+        whereArgs.add(sectorIdFilter);
+      }
+      query.setWhereCondition(whereClauses.join(' AND '), whereArgs);
     } else {
       _logger.debug('Retrieving complete summit list');
     }
+
     query.orderByColumns = <String>[SummitNamesTable.columnName];
 
     // Run the database query